+++ tests/testthat/test-ml-pipeline.R
@@ -1,4 +1,3 @@
-skip("Skip for now")
 skip_on_ci()
 skip_spark_min_version(3.5)